Some more refactoring and VHDL transformation

+++ apps/uscxml-browser.cpp
@@ -27,7 +27,7 @@
 #ifdef HAS_EXECINFO_H
 void printBacktrace(void** array, int size) {
 	char** messages = backtrace_symbols(array, size);
-	for (int i = 0; i < size && messages != NULL; ++i) {
+	for (size_t i = 0; i < size && messages != NULL; ++i) {
 		std::cerr << "\t" << messages[i] << std::endl;
 	}
 	std::cerr << std::endl;
